Removed parsing of bigint to number by default

+++ tests/setup.ts
@@ -76,10 +76,10 @@ export function prepareDatabase(conn: Connection): Promise<any> {
         {
             text: `SELECT * INTO TEMPORARY tmp_users
                 FROM (VALUES 
-		            (1::bigint,	'Irakliy'::VARCHAR, now()::timestamptz, now()::timestamptz),
-		            (2::bigint,	'Yason'::VARCHAR, 	now()::timestamptz, now()::timestamptz),
-		            (3::bigint,	'George'::VARCHAR, 	now()::timestamptz, now()::timestamptz),
-                    (4::bigint,	'T''est'::VARCHAR, 	now()::timestamptz, now()::timestamptz)
+		            (1::int, 'Irakliy'::VARCHAR,  now()::timestamptz, now()::timestamptz),
+		            (2::int, 'Yason'::VARCHAR, 	  now()::timestamptz, now()::timestamptz),
+		            (3::int, 'George'::VARCHAR,   now()::timestamptz, now()::timestamptz),
+                    (4::int, 'T''est'::VARCHAR,   now()::timestamptz, now()::timestamptz)
 	            ) AS q (id, username, created_on, updated_on);`
         }
     ]);
